+++ projects/np-ui-lib/src/lib/np-calendar/np-calendar.module.ts
@@ -1,0 +1,12 @@
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+import { NpCalendarComponent } from './np-calendar.component';
+
+@NgModule({
+  declarations: [NpCalendarComponent],
+  imports: [
+    CommonModule
+  ],
+  exports: [NpCalendarComponent]
+})
+export class NpCalendarModule { }